requirements.txt
boto3
pandas
pyarrow
snowflake-connector-python
python-dotenv
gspread
oauth2client
psycopg2-binary